+++ src/scripts/static/toolbox.gd
@@ -129,7 +129,7 @@ static func format_time_str_from_frame(frame_count: int) -> String:
 	return format_time_str(float(frame_count) / Project.get_framerate())
 	
 
-static func format_time_str(total_seconds: float) -> String:
+static func format_time_str(total_seconds: float, short: bool = false) -> String:
 	var total_seconds_int: int = floor(total_seconds)
 
 	var hours: int = int(float(total_seconds_int) / 3600)
@@ -138,7 +138,10 @@ static func format_time_str(total_seconds: float) -> String:
 	var seconds: int = total_seconds_int % 60
 	var micro: int = int(float(total_seconds - total_seconds_int) * 100)
 
-	return "%02d:%02d:%02d.%02d" % [hours, minutes, seconds, micro]
+	if short:
+		return "%02d:%02d:%02d" % [hours, minutes, seconds]
+	else:
+		return "%02d:%02d:%02d.%02d" % [hours, minutes, seconds, micro]
 	
 
 static func find_subfolder_files(files: PackedStringArray) -> PackedStringArray: